requirements.txt
asgiref==3.5.2
Django==4.0.6
django-phonenumber-field==6.3.0
mysqlclient==2.1.1
phonenumbers==8.12.51
sqlparse==0.4.2
